refactor sample_table
